refactor(ci): extract parseIntEnv to shared scripts/review/env.ts
Caught by the bot on its review of the previous fix commit: I had
literally copy-pasted parseIntEnv between index.ts and reply.ts. Real
DRY violation; the helper is small but the duplication signal stays
correct as the codebase grows.

Skipped a second finding from the same review pass — model ID
validation against an allowlist. The Anthropic SDK already throws a
clear error on unknown models, and an allowlist would lag behind new
model releases. Defense-in-depth not worth the maintenance cost here.
